Proposal: Replace wrapping_add with checked_add
`u16` seems uncomfortably small a generation size for me to be sure that I won't overflow it by accident if I have, I dunno, a particle system that adds and removes lots of objects rapidly.  It looks like you use `wrapping_add` for all the generation increments, so it appears that a generation that gets maxed out will just wrap around, possibly making stale handles look alive again.  This PR (hopefully) replaces them all with `checked_add()` followed by an `expect()`.  This makes it a dynamically-checked error for a generation number to overflow.

Whether you want this is up to you, I just wanted to offer the suggestion.